[EPO-386] Jupyter performance
I noted that downloading a particular layer of the lsstsqre/jld-lab image
was very very slow.  All our epo images had the same base image, even done
on different days.  I then noticed that docker build wasn't checking the
registry for new images, which we want.  So I'm adding an extra pull
before the build to make sure we get the latest image.  I'm also setting
the tag to latest in the Dockerfile, although this still won't trigger
a check of the remote registry.