[EPO-384] Use no-cache option when building daily images.
If we don't use no-cache, then we might not build the image if only
the notebooks (in another repo) changed.  So we should force the image
creation to be fresh and not use any caching.  If you want to use caching,
then copy paste the command and run it in your dev flow.