[EPO-546] Disable persistent storage
By default, the helm chart was making a k8s PVC for each user logging in.
This is trouble for us, and we don't need it anyway.  We'll be storing things
off to the side anyway.

This also gets the passwordless sudo working.  I'm assuming there was some kind
of problem with the conflation of persistent storage and the UID passed in that
was giving me the unwritable home dir problem that was keeping the container
from starting up.