[EPO-546] Create a latest tag for images, and have the hub follow it.
Before, we had the spawner which was presenting a list of images, and
following latest was a bad idea because we never knew what might happen
in each image.  Now that we have a more stable base image, we can
just track to that.  The image is also 75% smaller, so we should have less
problems if we need to download a new one while waiting to log in.