[EPO-686] Fix ingress rules to get virtual hostnames to work.
Apparently even though I'm just doing SSL/HTTPS, I have to provide
the rules in an HTTP format (I guess because the backend side is HTTP?),
even though the backend is everything with that hostname.

When I do it like this, everything seems to get to the right place.
Before, everything would go to whatever the first ingress rule was.