[EPO-604] Install hide_code from github, not npm.
The npm version doesn't seem to work as well as the github one, and
by installing it from github it will never be out of date.

Also, fix the permissions on the investigations directory, otherwise
people will get strange errors trying to save their notebooks.  It's
just a copy anyway.