[EPO-634] Question widget first attempt.
This wraps the logic to make a question box, and provides just one
call with the text and the number of rows for the answer box.  This
at least gets all the creation in one place (and out of the notebook).

This will help us later refactor the questions into saving the answers
to a remote database.

This also contains some CSS to make sure that the text wraps appropriately.
Thanks Abbey!