[EPO-879] Provide better origin URL for cross-origin checks.
Initially I just put in * as the origin url, which allows anyone to
connect.  This should really be the external url of the website,
such as epo-staging.lsst.rocks or sandbox-1.lsst.rocks.  This prevents
cross-origin scripting and other security issues.

Since we already have the external URL, return that for the case when
we are using the function.  For local running, localhost:8888 works fine.