[EPO-881] Try newer version of JupyterHub chart.
This change allows us to use a newer version of Jupyter.  There have
been a few bugs fixed in the kubespawner, which should be picked up
by this change.  This may make the issue of being unable to log in to
go away.

I've tested this out on a one time environment, and the widgets and
notebooks seem to be fine.

I've been unable on this environment, or a test environment hitting
some random URLs to reproduce the inability to log in, but it seems
to happen somewhat frequently on our normal environments.  Hopefully
more debug output will help with this as well to let us know more
when it happens again.