[EPO-838] Refactor stage 1: remove cluster query from widget.
First steps to a big refactor:

1. Remove the SDSS query from the Skyviewer + H-R diagram widget.
The widget shouldn't need to have the query or knowledge about its
data.

2. Take an astropixie.cluster object in the constructor to the
Skyviewer + H-R diagram widget.  This is the cluster to operate on,
and allows us to later pass in other clusters or create the clusters
based on the text box (although this isn't currently enabled).

3. Clean up usage of IDs in the Skyviewer + H-R diagram widget.
Instead of referring to objID and ID fields, which seem to be the same,
use the ids function that an astropixie.cluster provides, to provide
an abstraction for the IDs.

4. Stop swallowing exceptions.  There were try ... except blocks
in the Skyviewer + H-R diagram widget, which were really annoying.
Any errors would simply print out the message, but not the stack
trace.  The logging library would only print out the log line of
where it was caught, which was completely useless.  Bokeh's Tornado
server catches the exceptions anyway, so there was no need to do this,
and it was actually harmful in that it threw away valuable data.

5. Clean up the astropixie.OpenCluster object.  Use coord as part
of the interface, rather than ra/dec.  Make all clusters have a distance
property, which is used in calculations in the science.py.

6. Refactor the SDSSCluster to have the proper inheritance.  Berkely20
is a specific SDSS query, not the other way around.  Now we have an
SDSSRegion class that can take a query, which can easily be subclassed
with other properties to represent a cluster.

Still much to do.