[EPO-838] Refactor stage 2: Fix SDSS data.
1) Store the initial table returned by the SDSS query in a new
table property on the cluster object.  This is used by the skyviewer,
and is passed in raw, just like before.

2) Stop trying to mess around converting the tables, and just make
it into a numpy array by passing it to the numpy.array() constructor.

3) Use numpy for math, instead of for loops.  This is the same, and
much simpler and easier to deal with and understand.  The math operates
on each column at once, using referenced values for its row.  Magical.

4) We were using V, R, and V-R and passing them into the temperature
function assuming they were B and B-V.  This undoubtedly is a problem
and will cause us to have wrong temperature values, which we were thinking
we might have, even with the new equations.

5) Put a comment with a link to the relevant equations made by Robert
Lupton for future reference, and put it in the code next to where the
equations are used so that mistakes like this are less hidden in the future.