[EPO-838] Get rid of _add_null_selection, get the graph to display again.
There were issues both with the adding of IDs to the numpy array,
and the null selection is not the way this should be done.

Bokeh has problems rendering the graph if the ids are a maskable numpy
array, so convert it to a python array while creating the column data
source.

More work to come.