[EPO-838] Refactor stage 3: Clean up creation of H-R diagram, calculate temp up front.
1) Calculate the temperature and luminosity on creation of the widget,
not when showing the data.  Next will be not recalculating temperature
and luminosity when selections are made.

2) Update range creation based on #1.  Use the precalculated data
to create the ranges, and use numpy functions to find max and min.

3) Clean up creation of figures, axes, sliders.  Pass in more values
to the constructors instead of after they are created.  Break up each
parameter to a single line.

4) Use the same labels for axes, hover, and range sliders that all
include the same text and units.

5) Get rid of variables that are used just once (name, line_color)
and just pass them in instead of having a variable and passing in
the variable.  Having a variable implies that it is used in more than
one place, but these aren't.