[EPO-838] Refactor: Stage 4.  Rework callbacks.
1) Start off with a blank H-R diagram.

2) Make a new function, _filter_cluster_data that will look at the
ids selected by the Aladin lasso and filter the cluster data
down to just what is selected by the lasso by using np.isin and
then applying the mask.

3) Make the cluster data dictionary a bit more explicit, instead
of using x and y, use temperature and luminosity.  This is in the
data dictionary, the hover tool, and the call to pf.circle.  This
also changes the _filter_indices_on_sliders function when we look
through self.selected_data.

4) Store the ColumnDataSource, instead of what is returned by
pf.Circle.  This is the appropriate thing to do, and allows the
H-R diagram to be redrawn automatically if the data property is set
to a new data dictionary.

4) Fix hover tool.  By using the same ColumnDataSource, the hover
tool works even after a skyviewer selection or changes to the
contents of the H-R diagram.  This fixes EPO-830.

5) Change the _hr_selection function to look up the SDSS ids in
the selected data dictionary to highlight them back in the skyviewer.

6) Stop recomputing temperature and luminosity on _skyviewer_selection.
Instead use the precomputed self.selected_data.  This allows us to
delete the _filter_selection function.

7) Delete _set_selection_ids.  Don't try to use an ID of 0 (which doesn't
exist anyway) if selection_ids is empty.  It will end up empty normally
in the list comprehension and select nothing appropriately (with an
empty H-R diagram).

8) Clean up _skyviewer_selection.  Instead of looking up items based
on selecting the name and using the stored figure, just use the stored
columndatasource talked about in #4 above.  This gets rid of all the problems
with selected[0] and if the data source isn't found, since we always have
a reference to it.  It always exists by the time this callback is called,
since we store the reference before setting up the callback.  There's
no need for additional error checking.

Still much work, mostly around the sliders, filters, and selection logic,
and trying to wire up the reset button appropriately to reset all of those
items correctly when doing new selections.  But we're slowly getting there.