[EPO-838] Include sliders in filter_cluster_data.
Do the filtering based on the temperature and luminosity sliders
in the _filter_cluster_data function.  This means that even if the
sliders aren't shown, they are assumed for simplicity to exist, and
set to the full range.

We add these to the mask of items that we choose to display on the
H-R diagram.

For this reason, always create the sliders, even if they aren't
displayed.  Only have the show_sliders property control if we display
them, not create them.

Remove the _filter_indices_on_sliders function, this is now all
wrapped up in the _filter_cluster_data function.  Remove the call
in the _skyviewer_selection call, and stop trying to do the bokeh
selection on what is shown on the sliders.  Just display them or not
display them, but don't bother with the selection property anymore.