[EPO-838] Add reset functionality.
Add back the reset button, and make it:

1) Reset the sliders back to selecting everything.
2) Reset any selection lassos in the H-R diagram, and go back to what
was initially selected in the skyviewer (which we now store in
self.skyviewer_ids.

The only thing this doesn't reset is the selection in the skyviewer.
This allows for people to select ranges and lassos in the H-R diagram
without having to redraw their lasso in the skyviewer, so they are
playing with exactly the same data.