[EPO-980] Increase time to 5 seconds before loading catalog into skyviewer.
For the workshop, the last thing I want is for people to hit this
race condition.  It's not ideal, but I'd rather spend a few extra
seconds in loading time than to have to coach people in how to
restart the kernels and reload the notebooks.

This can be taken out when a better solution is implemented, which
I talked about in the associated ticket.